fixed autoduration bug and cut segments bug
for autoduration bug replaced map with foreach this makes sense?
i guess the value wasn't returned from the function properly inside
a map but inside a foreach it was ok?
for cut segments just rollbacked the editpoint func
it behaved extremely strangely and i'm still not sure why
but overall the edit point func is used weirdly with every edit calling
the function twice?
i'll need to return to this at a later date

signed off by IdoDePro
<idozipori08@gmail.com>